requirements.txt
gradio
requests
langchain
langchain-community
langchain-core
pgvector
python-dotenv
langchain-huggingface
langchain-groq
langchain-tavily    
langchain-chroma
langgraph
huggingface_hub
supabase
arxiv
pymupdf
wikipedia
transformers
sentence-transformers
tqdm
